Log entries and loading images from file.
Main change is that many methods now create log-entries so that the actions of the project can be followed by the user.

In addition the function that loads images from file has been changed so that by default intensity values as written in the file are read (rather than filtered data.) User can change settings to active filter.